BigDecimal, update cart price

+++ src/main/java/com/example/WebApp/dto/CartDto.java
@@ -6,13 +6,15 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.experimental.FieldDefaults;
 
+import java.math.BigDecimal;
+
 @Data
 @FieldDefaults(level = AccessLevel.PRIVATE)
 @AllArgsConstructor
 public class CartDto {
 
     //@NotNull(message = "Total price can't be null")
-    Long totalPrice;
+    BigDecimal totalPrice;
 
     @NotNull(message = "User id can't be null")
     Long userId;
